requirements.txt
fastmcp==2.12.5
mcp==1.14.0
pyjwt==2.10.1
python-dotenv>=1.1.0
cryptography==46.0.3
